requirements.txt
apache-airflow
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv